refactor(runtime-config): modularize internals with stable facade
